+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_redundant_async/visitor.dart
@@ -53,7 +53,9 @@ class _AsyncVisitor extends RecursiveAstVisitor<void> {
 
     final type = node.expression?.staticType;
 
-    if (type == null || !type.isDartAsyncFuture) {
+    if (type == null ||
+        !type.isDartAsyncFuture ||
+        type.nullabilitySuffix == NullabilitySuffix.question) {
       hasValidAsync = true;
     }
   }