+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_redundant_async/examples/example.dart
@@ -30,4 +30,8 @@ class SomeClass {
 
     print(records);
   }
+
+  Future<void> returnNullable(SomeClass? instance) async {
+    return instance?.report([]);
+  }
 }